+++ bolt09/index.js
@@ -1,0 +1,9 @@
+const featureFlagDetails = require('./feature_flag_details');
+const featureFlagsAsWords = require('./feature_flags_as_words');
+const featureFlagsFromWords = require('./feature_flags_from_words');
+
+module.exports = {
+  featureFlagDetails,
+  featureFlagsAsWords,
+  featureFlagsFromWords,
+};